[refactor] 공통 패키지를 infrastructure layer로 이동
