+++ src/main/java/junseok/snr/wallet/application/service/TransactionService.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.application.service;
 
 import jakarta.persistence.criteria.Predicate;
-import junseok.snr.wallet.common.Web3jUtils;
+import junseok.snr.wallet.infrastructure.repository.common.Web3jUtils;
 import junseok.snr.wallet.api.controller.dto.WithdrawDto;
 import junseok.snr.wallet.domain.Transaction;
 import junseok.snr.wallet.domain.TransactionType;
