+++ src/main/java/junseok/snr/wallet/domain/Wallet.java
@@ -2,7 +2,7 @@ package junseok.snr.wallet.domain;
 
 
 import jakarta.persistence.*;
-import junseok.snr.wallet.common.AESEncryptionService;
+import junseok.snr.wallet.infrastructure.repository.common.AESEncryptionService;
 import junseok.snr.wallet.application.service.ExceptionCode;
 import junseok.snr.wallet.application.service.WalletException;
 import lombok.*;
