+++ src/main/java/junseok/snr/wallet/infrastructure/repository/common/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.common;
+package junseok.snr.wallet.infrastructure.repository.common;
 
 import junseok.snr.wallet.application.service.ExceptionCode;
 import lombok.Builder;
